# Conflicts:
#	.idea/workspace.xml

+++ src/main/java/com/roomster/roomsterbackend/service/impl/CommentPostService.java
@@ -2,14 +2,16 @@ package com.roomster.roomsterbackend.service.impl;
 
 import com.roomster.roomsterbackend.dto.BaseResponse;
 import com.roomster.roomsterbackend.dto.comment.CommentPostDto;
+import com.roomster.roomsterbackend.dto.user.PartUser;
 import com.roomster.roomsterbackend.entity.CommentEnity;
+import com.roomster.roomsterbackend.entity.PostEntity;
 import com.roomster.roomsterbackend.entity.UserEntity;
 import com.roomster.roomsterbackend.mapper.CommentMapper;
 import com.roomster.roomsterbackend.repository.CommentPostRepository;
 import com.roomster.roomsterbackend.repository.PostRepository;
+import com.roomster.roomsterbackend.repository.RoleRepository;
 import com.roomster.roomsterbackend.repository.UserRepository;
 import com.roomster.roomsterbackend.service.IService.ICommentPostService;
-import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.stereotype.Service;
@@ -18,7 +20,6 @@ import org.springframework.transaction.annotation.Transactional;
 import java.security.Principal;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -31,6 +32,8 @@ public class CommentPostService implements ICommentPostService {
     PostRepository postRepository;
     @Autowired
     CommentMapper commentMapper;
+    @Autowired
+    RoleRepository roleRepository;
 
     @Override
     public CommentPostDto saveNewComment(CommentPostDto commentPostDTO, Principal connectedUser) {
@@ -41,30 +44,54 @@ public class CommentPostService implements ICommentPostService {
     }
 
     @Override
-    public CommentPostDto updateComment(Long commentId,CommentPostDto commentPostDTO) {
+    public CommentPostDto updateComment(Long commentId, CommentPostDto commentPostDTO, Principal connectedUser) {
         Optional<CommentEnity> oldComment = commentPostRepository.findById(commentId);
-        if(oldComment.isPresent()){
-            oldComment.get().setTitle(commentPostDTO.getTitle());
-            oldComment.get().setContent(commentPostDTO.getContent());
-            return commentMapper.entityToDTO(commentPostRepository.save(oldComment.get()));
+        var user = (UserEntity)((UsernamePasswordAuthenticationToken)connectedUser).getPrincipal();
+        if (oldComment.isPresent() && user != null) {
+            if(user.getId().equals(oldComment.get().getUserId())) {
+                oldComment.get().setContent(commentPostDTO.getContent());
+                commentPostRepository.save(oldComment.get());
+                return commentMapper.entityToDTO(oldComment.get());
+            }
         }
         return null;
     }
 
     @Override
-    public BaseResponse deleteComment(Long commentId) {
+    public BaseResponse deleteComment(Long commentId, Principal connectedUser) {
+        var user = (UserEntity)((UsernamePasswordAuthenticationToken)connectedUser).getPrincipal();
+        Optional<CommentEnity> comment = commentPostRepository.findById(commentId);
         try {
-            commentPostRepository.deleteById(commentId);
-        }catch (Exception ex){
+            if(comment.isPresent() && user != null){
+               if(comment.get().getUserId().equals(user.getId())){
+                   commentPostRepository.deleteById(commentId);
+                   return BaseResponse.success("Bạn đã xóa bình luận!");
+               }
+            }
+        } catch (Exception ex) {
             BaseResponse.error(ex.getMessage());
         }
-        return BaseResponse.success("Xóa Thành Công!");
+        return BaseResponse.error("Không thể xóa bình luận");
     }
 
     @Override
     public List<CommentPostDto> getAllCommentOfPost(Long postId) {
-        return commentPostRepository.getCommentByPostId(postId)
-                .stream().map(commentPostEntity -> commentMapper.entityToDTO(commentPostEntity))
-                .collect(Collectors.toList());
+        List<CommentPostDto> commentPostDtos = commentPostRepository.getCommentByPostId(postId).stream().map(commentEnity -> commentMapper.entityToDTO(commentEnity)).toList();
+
+        for (CommentPostDto item : commentPostDtos
+        ) {
+            PartUser partUser = new PartUser();
+            UserEntity user = userRepository.findById(item.getUserId()).orElseThrow();
+            partUser.setUserName(user.getUserName());
+            partUser.setImages(user.getImages());
+            Optional<PostEntity> post = postRepository.findById(item.getPostId());
+            if(post.isPresent()) {
+                if (item.getUserId().equals(post.get().getCreatedBy())) {
+                    partUser.setUserName("Tác Giả");
+                }
+            }
+            item.setPartUser(partUser);
+        }
+        return commentPostDtos;
     }
 }
